requirements.txt
accelerate==0.31.0
aiofiles==22.1.0
aiohttp==3.9.5
aiosignal==1.3.1
aiosqlite==0.18.0
anyio==3.5.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astroid==2.15.8
asttokens==2.0.5
async-timeout==4.0.3
attrs==22.1.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==4.1.0
Bottleneck==1.3.5
brotlipy==0.7.0
certifi==2024.7.4
cffi==1.15.1
charset-normalizer==2.0.4
comm==0.1.2
continuum==1.2.7
contourpy==1.0.5
cryptography==39.0.1
cssutils==2.11.1
cycler==0.11.0
Cython==0.29.35
dataframe-image==0.2.3
datasets==2.19.2
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
dodgy==0.2.1
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.2.1
executing==0.8.3
fastjsonschema==2.16.2
filelock==3.9.0
flake8==5.0.4
flake8-polyfill==1.0.2
fonttools==4.25.0
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
gmpy2==2.1.2
h5py==3.11.0
html2image==2.0.4.3
huggingface-hub==0.23.2
idna==3.4
ImageHash==4.3.1
imageio==2.34.1
importlib-metadata==6.0.0
importlib-resources==5.2.0
iniconfig==2.0.0
ipykernel==6.19.2
ipython==8.12.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isort==5.13.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.6
jsonschema==4.17.3
jupyter==1.0.0
jupyter_client==8.1.0
jupyter-console==6.6.3
jupyter_core==5.3.0
jupyter-events==0.6.3
jupyter_server==2.5.0
jupyter_server_fileid==0.9.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.8.0
jupyter-ydoc==0.2.4
jupyterlab==3.6.3
jupyterlab-pygments==0.1.2
jupyterlab_server==2.22.0
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
lazy_loader==0.4
lazy-object-proxy==1.10.0
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==0.8.4
mkl-fft==1.3.6
mkl-random==1.2.2
mkl-service==2.4.0
more-itertools==10.2.0
mpmath==1.2.1
multidict==6.0.5
multiprocess==0.70.16
munkres==1.1.4
mypy==1.10.0
mypy-extensions==1.0.0
nbclassic==0.5.5
nbclient==0.5.13
nbconvert==6.5.4
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==2.8.4
notebook==6.5.4
notebook_shim==0.2.2
numexpr==2.8.4
numpy==1.24.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
openpyxl==3.1.5
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
peft==0.9.0
pep8-naming==0.10.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pip==23.1.2
pkgutil_resolve_name==1.3.10
platformdirs==2.5.2
pluggy==1.5.0
ply==3.11
pooch==1.4.0
POT==0.9.0
prometheus-client==0.14.1
prompt-toolkit==3.0.36
prospector==1.10.3
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pyarrow-hotfix==0.6
pycodestyle==2.9.1
pycparser==2.21
pydocstyle==6.3.0
pyflakes==2.5.0
Pygments==2.15.1
pylint==2.17.7
pylint-celery==0.3
pylint-django==2.5.3
pylint-flask==0.6
pylint-plugin-utils==0.7
pyOpenSSL==23.0.0
pyparsing==3.0.9
PyQt5-sip==12.11.0
pyrsistent==0.18.0
PySocks==1.7.1
pytest==8.2.2
pytest-mock==3.14.0
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2022.7
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.1.0
qtconsole==5.4.2
QtPy==2.2.0
quadprog==0.1.11
regex==2024.5.15
requests==2.32.3
requirements-detector==1.2.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
safetensors==0.4.3
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.13.2
semver==3.0.2
Send2Trash==1.8.0
setoptconf-tmp==0.3.1
setuptools==67.8.0
sip==6.6.2
six==1.16.0
smmap==5.0.1
sniffio==1.2.0
snowballstemmer==2.2.0
soupsieve==2.4
stack-data==0.2.0
sympy==1.11.1
terminado==0.17.1
threadpoolctl==2.2.0
tifffile==2023.7.10
tinycss2==1.2.1
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.5
torch==2.3.0
torchvision==0.18.0
tornado==6.2
tqdm==4.65.0
traitlets==5.7.1
transformers==4.41.2
triton==2.3.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==1.26.16
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.38.4
widgetsnbextension==4.0.5
wrapt==1.16.0
xxhash==3.4.1
y-py==0.5.9
yarl==1.9.4
ypy-websocket==0.8.2
zipp==3.11.0
